RStudio codes for assessing a .csv file
This file contain the codes to import a .csv file to RStudio, read the file data and rearranging them accordingly. Furthermore, it contain the codes to make subsets of data, perform certain mathematical operations such as calculating the mean values and standard deviations of sets of data and obtain boxplots.